Rename to vscode-button-group and remove the divider tag in favour of adding automatically
