+++ src/vscode-button-group/index.ts
@@ -1,0 +1,1 @@
+export {VscodeButtonGroup} from './vscode-button-group.js';
