--- src/vscode-split-button/index.ts
@@ -1,1 +1,0 @@
-export {VscodeSplitButton} from './vscode-split-button.js';
